Better warning supression
