+++ corebehrt/azure/util/azure.py
@@ -1,4 +1,4 @@
-import sys
+from corebehrt.azure.util.supress import supress_warnings
 
 AZURE_AVAILABLE = False
 
@@ -6,8 +6,9 @@ try:
     #
     # Check if azure is available and set flag.
     #
-    from azure.ai.ml import MLClient
-    from azure.identity import DefaultAzureCredential
+    with supress_warnings():
+        from azure.ai.ml import MLClient
+        from azure.identity import DefaultAzureCredential
 
     AZURE_AVAILABLE = True
 except:
@@ -38,38 +39,3 @@ def ml_client() -> "MLClient":
     """
     check_azure()
     return MLClient.from_config(DefaultAzureCredential())
-
-
-class AzureWarningSupressor:
-    def __init__(self, patterns):
-        self.patterns = patterns
-
-    def __enter__(self):
-        sys.stdout = AzurePrintFilter(sys.stdout, self.patterns)
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        sys.stdout = sys.stdout.stream
-
-
-class AzurePrintFilter(object):
-    def __init__(self, stream, patterns):
-        self.stream = stream
-        self.patterns = patterns
-        self.triggered = False
-
-    def __getattr__(self, attr_name):
-        return getattr(self.stream, attr_name)
-
-    def write(self, data):
-        if data == "\n" and self.triggered:
-            self.triggered = False
-        else:
-            for pattern in self.patterns:
-                if pattern in data:
-                    self.triggered = True
-                    return
-            self.stream.write(data)
-            self.stream.flush()
-
-    def flush(self):
-        self.stream.flush()
