+++ corebehrt/azure/util/supress.py
@@ -1,0 +1,30 @@
+import sys
+import contextlib
+
+SUPRESS_WARNINGS = [
+    "pathOnCompute is not a known attribute of class",
+    ".*This is an experimental class.*",
+    "Found the config file in: /config.json",
+]
+
+
+@contextlib.contextmanager
+def suppress_warnings():
+    stdout_backup = sys.stdout
+    sys.stdout = AzureWarningFilter(sys.stdout)
+    yield
+    sys.stdout = stdout_backup
+
+
+class AzureWarningFilter(object):
+    def __init__(self, stream):
+        self.stream = stream
+
+    def __getattr__(self, attr_name):
+        return getattr(self.stream, attr_name)
+
+    def write(self, data):
+        for pattern in SUPRESS_WARNINGS:
+            if pattern in data:
+                return
+        self.stream.write(data)
